// StudentWeb/src/app/models/Student.ts
export class Student{
    id?: number;
    userName: string;
    firstName: string;
    lastName: string;
    age: number;
    career: string;

    constructor(){         
        this.userName = "" ;
        this.firstName = "";
        this.lastName = "";
        this.age = 0;
        this.career = "";
    }
}